fix(backend-sqlite): address PR-378 bot-review findings
Copilot:
- split_exec_id now enforces the u16 partition-index invariant
  explicitly; out-of-range partitions surface as Validation(InvalidInput)
  rather than silently flowing into an i64 partition_key bind.
- suspend_inner guards against empty SuspendArgs::waitpoints up-front
  (with_waitpoints(vec![]) can empty the vec post-construction).
- seed + rotate now share validate_kid + validate_secret_hex helpers,
  rejecting empty or colon-containing kid and enforcing the 64-char
  (256-bit) secret_hex invariant on both paths. Rotate previously only
  checked hex-parse success.
- Duplicated begin_immediate / commit_or_rollback / rollback_quiet /
  now_ms / split_exec_id helpers extracted to crates/ff-backend-sqlite/
  src/tx_util.rs so backend.rs + suspend_ops.rs share one copy.

Gemini:
- DistinctSignals counter uses HashSet<Uuid> instead of
  HashSet<String>, dropping the per-signal String allocation.
- INSERT_SIGNAL_EVENT_SQL rustdoc enumerates all 7 binds + clarifies
  that bind 7 is a BLOB lookup against ff_exec_core while bind 1 is
  the TEXT stringification written into ff_signal_event. Gemini's
  HIGH priority string-vs-BLOB findings at suspend_ops.rs:973/1061
  were false positives — ff_signal_event.execution_id and .signal_id
  are TEXT columns, not BLOB, which matches the PG reference schema.

ff-core::crypto::hmac:
- hmac_sign rustdoc now documents the kid-shape constraint ('kid'
  must not contain ':') with a pointer to the backend validators.
  The primitive does not re-validate for performance; seed + rotate
  entry points on both backends already reject colon kids.

Tests still green: 13 SqliteBackend integration tests + 5 ff-core
crypto::hmac unit tests + 20 PG lib tests.
