+++ crates/ff-backend-sqlite/src/lib.rs
@@ -35,6 +35,7 @@ pub mod queries;
 mod registry;
 pub mod retry;
 mod suspend_ops;
+mod tx_util;
 
 pub use backend::SqliteBackend;
 pub use errors::{is_retryable_sqlite_busy, MAX_ATTEMPTS};
